+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [1.1.4]
+
+### Bug Fixes
+
+- context precedence ([#27](https://github.com/rudderlabs/rudder-sdk-rust/pull/27))
+- context override in batch ([#25](https://github.com/rudderlabs/rudder-sdk-rust/pull/25))
+
 ## [1.1.3]
 
 ### Bug Fixes